# Faster, more accurate Poisson Binomial PMF